v1.1: change failure handling for exporter/telegraf
This release substantially changes the behavior of the exporter
and telegraf containers in the pod: previously, a failure in either
of these processes would lead to the container failing and being
restarted by the kubelet, potentially leading to the entire pod being
marked as unavailable.  Now by default an error is logged and the
process is restarted by the wrapper script, but the container does
not exit.

An `exitOnError` helm option has been added to both the `exporter`
and `telegraf` config value sections: set each to "true" to restore
the old behavior.