v1.4.0: remove telegraf
Telegraf was a temporary workaround for the fact that GKE originally
had no simple method of collecting application-level metrics.

Google's Managed Service for Prometheus basically solves this problem
and is on by default in new GKE clusters; there's no reason for us
to track and redistribute upstream telegraf any more.